chore: remove a/b test for defi referrals
