+++ ui/pages/core/defi-referral-consent/defi-referral-consent.test.tsx
@@ -5,17 +5,9 @@ import { renderWithProvider } from '../../../../test/lib/render-helpers-navigate
 import mockState from '../../../../test/data/mock-state.json';
 import { DEFI_REFERRAL_PARTNERS } from '../../../../shared/constants/defi-referrals';
 import { enLocale as messages } from '../../../../test/lib/i18n-helpers';
-import { useABTest } from '../../../hooks/useABTest';
-import {
-  DefiReferralUIABTestVariant,
-  DEFI_REFERRAL_CONSENT_AB_TEST_VARIANTS,
-} from '../../../../shared/lib/ab-testing/configs/defi-referral-ui';
 import { DefiReferralConsent } from './defi-referral-consent';
 
-jest.mock('../../../hooks/useABTest');
-
 const mockStore = configureMockStore([]);
-const mockUseABTest = jest.mocked(useABTest);
 
 // Get all partners as test cases
 type PartnerTestCase = {
@@ -35,11 +27,6 @@ const partnerTestCases: PartnerTestCase[] = Object.values(
 describe('DefiReferralConsent', () => {
   beforeEach(() => {
     jest.clearAllMocks();
-    mockUseABTest.mockReturnValue({
-      variant: DEFI_REFERRAL_CONSENT_AB_TEST_VARIANTS.control,
-      variantName: DefiReferralUIABTestVariant.Control,
-      isActive: false,
-    });
   });
 
   // @ts-expect-error This function is missing from the Mocha type definitions
@@ -54,7 +41,7 @@ describe('DefiReferralConsent', () => {
         learnMoreUrl,
       };
 
-      it('renders control variant content with checkbox flow', () => {
+      it('renders the component with correct title and subtitle', () => {
         const store = mockStore(mockState);
 
         renderWithProvider(<DefiReferralConsent {...props} />, store);
@@ -64,9 +51,8 @@ describe('DefiReferralConsent', () => {
             messages.defiReferralTitle.message.replace('$1', partnerName),
           ),
         ).toBeInTheDocument();
-        expect(screen.getByRole('checkbox')).toBeChecked();
         expect(
-          screen.getByRole('button', { name: messages.confirm.message }),
+          screen.getByText('MetaMask referral code', { exact: false }),
         ).toBeInTheDocument();
       });
 
@@ -83,79 +69,28 @@ describe('DefiReferralConsent', () => {
         );
       });
 
-      it('renders control variant link with correct URL', () => {
+      it('renders the learn more link with correct URL', () => {
         const store = mockStore(mockState);
 
         renderWithProvider(<DefiReferralConsent {...props} />, store);
 
-        const learnMoreLink = screen.getByRole('link', {
-          name: `${messages.learnMoreUpperCase.message}.`,
-        });
+        const learnMoreLink = screen.getByRole('link');
         expect(learnMoreLink).toHaveAttribute('href', learnMoreUrl);
         expect(learnMoreLink).toHaveAttribute('target', '_blank');
         expect(learnMoreLink).toHaveAttribute('rel', 'noopener noreferrer');
       });
 
-      it('submits approved false in control when checkbox is unchecked', () => {
-        const store = mockStore(mockState);
-        const mockOnActionComplete = jest.fn();
-
-        renderWithProvider(
-          <DefiReferralConsent
-            {...props}
-            onActionComplete={mockOnActionComplete}
-          />,
-          store,
-        );
-
-        fireEvent.click(screen.getByRole('checkbox'));
-        fireEvent.click(
-          screen.getByRole('button', { name: messages.confirm.message }),
-        );
-
-        expect(mockOnActionComplete).toHaveBeenCalledWith({
-          approved: false,
-          selectedAddress: '0x123',
-        });
-      });
-
-      it('renders treatment variant content with dual actions', () => {
-        mockUseABTest.mockReturnValue({
-          variant: DEFI_REFERRAL_CONSENT_AB_TEST_VARIANTS.treatment,
-          variantName: DefiReferralUIABTestVariant.Treatment,
-          isActive: true,
-        });
+      it('renders the checkbox with default checked state', () => {
         const store = mockStore(mockState);
 
         renderWithProvider(<DefiReferralConsent {...props} />, store);
 
-        const titleKey = `${partnerId}ReferralTitle` as keyof typeof messages;
-        const confirmKey =
-          `${partnerId}ReferralConfirmText` as keyof typeof messages;
-        expect(
-          screen.getByText(messages[titleKey].message),
-        ).toBeInTheDocument();
-        expect(
-          screen.getByRole('link', {
-            name: messages.defiReferralTerms.message,
-          }),
-        ).toHaveAttribute('href', learnMoreUrl);
-        expect(
-          screen.getByRole('button', { name: messages[confirmKey].message }),
-        ).toBeInTheDocument();
-        expect(
-          screen.getByRole('button', {
-            name: messages.defiReferralNoThanks.message,
-          }),
-        ).toBeInTheDocument();
+        const checkbox = screen.getByRole('checkbox');
+        expect(checkbox).toBeInTheDocument();
+        expect(checkbox).toBeChecked();
       });
 
-      it('submits approved true in treatment when confirm is clicked', () => {
-        mockUseABTest.mockReturnValue({
-          variant: DEFI_REFERRAL_CONSENT_AB_TEST_VARIANTS.treatment,
-          variantName: DefiReferralUIABTestVariant.Treatment,
-          isActive: true,
-        });
+      it('calls onActionComplete with approved=true when confirm is clicked and checkbox is checked', () => {
         const store = mockStore(mockState);
         const mockOnActionComplete = jest.fn();
 
@@ -167,10 +102,8 @@ describe('DefiReferralConsent', () => {
           store,
         );
 
-        const confirmKey =
-          `${partnerId}ReferralConfirmText` as keyof typeof messages;
         const confirmButton = screen.getByRole('button', {
-          name: messages[confirmKey].message,
+          name: messages.confirm.message,
         });
         fireEvent.click(confirmButton);
 
@@ -180,12 +113,7 @@ describe('DefiReferralConsent', () => {
         });
       });
 
-      it('submits approved false in treatment when no thanks is clicked', () => {
-        mockUseABTest.mockReturnValue({
-          variant: DEFI_REFERRAL_CONSENT_AB_TEST_VARIANTS.treatment,
-          variantName: DefiReferralUIABTestVariant.Treatment,
-          isActive: true,
-        });
+      it('calls onActionComplete with approved=false when confirm is clicked and checkbox is unchecked', () => {
         const store = mockStore(mockState);
         const mockOnActionComplete = jest.fn();
 
@@ -197,15 +125,29 @@ describe('DefiReferralConsent', () => {
           store,
         );
 
-        const cancelButton = screen.getByRole('button', {
-          name: messages.defiReferralNoThanks.message,
+        // Uncheck the checkbox first
+        const checkbox = screen.getByRole('checkbox');
+        fireEvent.click(checkbox);
+
+        const confirmButton = screen.getByRole('button', {
+          name: messages.confirm.message,
         });
-        fireEvent.click(cancelButton);
+        fireEvent.click(confirmButton);
 
         expect(mockOnActionComplete).toHaveBeenCalledWith({
           approved: false,
           selectedAddress: '0x123',
         });
+      });
+
+      it('renders the checkbox label', () => {
+        const store = mockStore(mockState);
+
+        renderWithProvider(<DefiReferralConsent {...props} />, store);
+
+        expect(
+          screen.getByText(messages.defiReferralCheckboxLabel.message),
+        ).toBeInTheDocument();
       });
     },
   );
