--- ui/pages/core/defi-referral-consent/defi-referral-consent.types.ts
@@ -1,10 +1,0 @@
-export type DefiReferralConsentProps = {
-  onActionComplete: (result: {
-    approved: boolean;
-    selectedAddress: string;
-  }) => void;
-  selectedAddress: string;
-  partnerId: string;
-  partnerName: string;
-  learnMoreUrl: string;
-};
