+++ ui/pages/core/defi-referral-consent/index.ts
@@ -1,2 +1,2 @@
 export { DefiReferralConsent } from './defi-referral-consent';
-export type { DefiReferralConsentProps } from './defi-referral-consent.types';
+export type { DefiReferralConsentProps } from './defi-referral-consent';
